Add ProcessStream
